backend/requirements.txt
django==2.2
djangorestframework==3.11
djangorestframework-simplejwt==4.3.0
django-filter==2.2.0
django-cors-headers==3.2.1